+++ src/main/java/com/github/os72/protocjar/maven/ProtocJarMojo.java
@@ -25,17 +25,26 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
+import java.util.Properties;
+
+import kr.motd.maven.os.Detector;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.io.filefilter.IOFileFilter;
 import org.apache.commons.io.filefilter.TrueFileFilter;
+import org.apache.maven.artifact.Artifact;
+import org.apache.maven.artifact.factory.ArtifactFactory;
+import org.apache.maven.artifact.repository.ArtifactRepository;
+import org.apache.maven.artifact.resolver.ArtifactResolver;
+import org.apache.maven.artifact.versioning.VersionRange;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.project.MavenProject;
 import org.sonatype.plexus.build.incremental.BuildContext;
 
 import com.github.os72.protocjar.Protoc;
+import com.github.os72.protocjar.ProtocVersion;
 
 /**
  * Compiles .proto files using protoc-jar embedded protoc compiler (external protoc executable also supported)
@@ -48,20 +57,8 @@ public class ProtocJarMojo extends AbstractMojo
 {
 	private static final String DEFAULT_INPUT_DIR = "/src/main/protobuf/".replace('/', File.separatorChar);
 
-	/**
-	 * The Maven project.
-	 * 
-	 * @parameter property="project"
-	 * @required
-	 * @readonly
-	 */
-	private MavenProject project;
-
-	/** @component */
-	private BuildContext buildContext;
-
-	/**
-	 * Specifies protoc version.
+    /**
+	 * Specifies the protoc version.
 	 * 
 	 * @parameter property="protocVersion"
 	 */
@@ -123,13 +120,23 @@ public class ProtocJarMojo extends AbstractMojo
 	private boolean cleanOutputFolder;
 
 	/**
-	 * Path to protoc plugin that generates code for the specified {@link #type}.
+	 * Path to the protoc plugin that generates code for the specified {@link #type}.
 	 * <p>
 	 * Ignored when {@code <outputTargets>} is given
 	 *
 	 * @parameter property="pluginPath"
 	 */
 	private String pluginPath;
+
+	/**
+	 * Maven artifact coordinates of the protoc plugin that generates code for the specified {@link #type}.
+	 * Format: "groupId:artifactId:version" (eg, "io.grpc:protoc-gen-grpc-java:1.0.1")
+	 * <p>
+	 * Ignored when {@code <outputTargets>} is given
+	 *
+	 * @parameter property="pluginArtifact"
+	 */
+	private String pluginArtifact;
 
 	/**
 	 * Output directory for the generated java files. Defaults to
@@ -213,7 +220,45 @@ public class ProtocJarMojo extends AbstractMojo
 	 */
 	private String protocCommand;
 
-	public void execute() throws MojoExecutionException {
+	/**
+	 * Maven artifact coordinates of the protoc binary to use
+	 * Format: "groupId:artifactId:version" (eg, "com.google.protobuf:protoc:3.1.0")
+	 *
+	 * @parameter property="protocArtifact"
+	 */
+	private String protocArtifact;
+
+	/**
+	 * The Maven project.
+	 * 
+	 * @parameter property="project"
+	 * @readonly
+	 * @required
+	 */
+	private MavenProject project;
+
+	/** 
+	 * @parameter default-value="${localRepository}" 
+	 * @readonly
+	 * @required
+	 */
+	private ArtifactRepository localRepository;
+
+	/** 
+	 * @parameter default-value="${project.remoteArtifactRepositories}" 
+	 * @readonly
+	 * @required
+	 */
+	private List<ArtifactRepository> remoteRepositories;
+
+	/** @component */
+	private BuildContext buildContext;
+	/** @component */
+	private ArtifactFactory artifactFactory;
+	/** @component */
+	private ArtifactResolver artifactResolver;
+
+    public void execute() throws MojoExecutionException {
 		if (project.getPackaging() != null && "pom".equals(project.getPackaging().toLowerCase())) {
 			getLog().info("Skipping 'pom' packaged project");
 			return;
@@ -225,12 +270,13 @@ public class ProtocJarMojo extends AbstractMojo
 			target.addSources = addSources;
 			target.cleanOutputFolder = cleanOutputFolder;
 			target.pluginPath = pluginPath;
+			target.pluginArtifact = pluginArtifact;
 			target.outputDirectory = outputDirectory;
 			target.outputDirectorySuffix = outputDirectorySuffix;
 			target.outputOptions = outputOptions;
 			outputTargets = new OutputTarget[] {target};
 		}
-
+		
 		for (OutputTarget target : outputTargets) {
 			target.addSources = target.addSources.toLowerCase().trim();
 			if ("true".equals(target.addSources)) target.addSources = "main";
@@ -243,21 +289,71 @@ public class ProtocJarMojo extends AbstractMojo
 			if (target.outputDirectorySuffix != null) {
 				target.outputDirectory = new File(target.outputDirectory, target.outputDirectorySuffix);
 			}
+
+			// Copy files to a temporary directory.
+			target.outputDirectoryFinal = target.outputDirectory;
+			target.outputDirectory = new File(target.outputDirectory.getParent(), target.outputDirectory.getName() + "-tmp");
 		}
 		
 		performProtoCompilation();
 	}
 
+	private void copyAndUpdateOutputTarget(OutputTarget outputTarget) throws MojoExecutionException {
+		File src = outputTarget.outputDirectory;
+		File dst = outputTarget.outputDirectoryFinal;
+
+		dst.mkdirs();
+
+		// List files in dst and copy them to target.
+		Collection<File> files = org.apache.commons.io.FileUtils.listFiles(src, null, true);
+
+		for (File generatedFile : files) {
+			String path = generatedFile.getAbsolutePath();
+			path = path.replaceAll("^" + src.getAbsolutePath(), dst.getAbsolutePath());
+			File targetFile = new File(path);
+
+			try {
+				if (org.apache.commons.io.FileUtils.contentEquals(generatedFile, targetFile)) {
+					getLog().debug("Skipping " + targetFile.getName());
+					continue;
+				}
+			} catch (IOException ignored) {
+ 				// Copy the file on any error, for e.g., if the dstFile does not exist.
+  			}
+
+			try {
+				// Remove the target file if it already exists.
+				targetFile.delete();
+				org.apache.commons.io.FileUtils.copyFile(generatedFile, targetFile);
+				getLog().debug("Updating " + targetFile.getName());
+			} catch (IOException e) {
+				throw new MojoExecutionException("Error copying file from " + targetFile + " to " + generatedFile, e);
+			}
+		}
+
+		// Update the output directory and restore to its true, final, path.
+		outputTarget.outputDirectory = outputTarget.outputDirectoryFinal;
+	}
+
 	private void performProtoCompilation() throws MojoExecutionException {
-		getLog().info("Protoc version: " + protocVersion);
-		if (protocCommand == null) {
+		String protocTemp = null;
+		if ((protocCommand == null && protocArtifact == null) || includeStdTypes) {
+			if (protocVersion == null || protocVersion.length() < 1) protocVersion = ProtocVersion.PROTOC_VERSION;
+			getLog().info("Protoc version: " + protocVersion);
+			
 			try {
 				File protocFile = Protoc.extractProtoc(protocVersion.replace(".", ""), includeStdTypes);
-				protocCommand = protocFile.getAbsolutePath();
+				protocTemp = protocFile.getAbsolutePath();
 			}
 			catch (IOException e) {
 				throw new MojoExecutionException("Error extracting protoc for version " + protocVersion, e);
 			}
+			
+			if (protocCommand == null && protocArtifact == null) protocCommand = protocTemp;
+		}
+		
+		if (protocCommand == null && protocArtifact != null) {
+			protocCommand = resolveArtifact(protocArtifact).getAbsolutePath();
 		}
 		getLog().info("Protoc command: " + protocCommand);
 		
@@ -269,7 +365,7 @@ public class ProtocJarMojo extends AbstractMojo
 		for (File input : inputDirectories) getLog().info("    " + input);
 		
 		if (includeStdTypes) {
-			File stdTypeDir = new File(new File(protocCommand).getParentFile().getParentFile(), "include");
+			File stdTypeDir = new File(new File(protocTemp).getParentFile().getParentFile(), "include");
 			if (includeDirectories != null && includeDirectories.length > 0) {
 				List<File> includeDirList = new ArrayList<File>();
 				includeDirList.add(stdTypeDir);
@@ -293,6 +389,10 @@ public class ProtocJarMojo extends AbstractMojo
 	}
 
 	private void preprocessTarget(OutputTarget target) throws MojoExecutionException {
+		if (target.pluginArtifact != null && target.pluginArtifact.length() > 0) {
+			target.pluginPath = resolveArtifact(target.pluginArtifact).getAbsolutePath();
+		}
+		
 		File f = target.outputDirectory;
 		if (!f.exists()) {
 			getLog().info(f + " does not exist. Creating...");
@@ -323,7 +423,7 @@ public class ProtocJarMojo extends AbstractMojo
 			if (input == null) continue;
 			
 			if (input.exists() && input.isDirectory()) {
-				Collection<File> protoFiles = FileUtils.listFiles(input, fileFilter, TrueFileFilter.INSTANCE);
+				Collection<File> protoFiles = FileUtils.listFiles(input, fileFilter, null);
 				for (File protoFile : protoFiles) {
 					if (target.cleanOutputFolder || buildContext.hasDelta(protoFile.getPath())) {
 						processFile(protoFile, protocVersion, targetType, target.pluginPath, target.outputDirectory, target.outputOptions);
@@ -353,11 +453,11 @@ public class ProtocJarMojo extends AbstractMojo
 		boolean testAddSources = "test".endsWith(target.addSources);
 		
 		if (mainAddSources) {
-			getLog().info("Adding generated classes to classpath");
+			getLog().info("Adding generated classes in " + target.outputDirectory.getAbsolutePath() + " to classpath");
 			project.addCompileSourceRoot(target.outputDirectory.getAbsolutePath());
 		}
 		if (testAddSources) {
-			getLog().info("Adding generated classes to test classpath");
+			getLog().info("Adding generated classes in " + target.outputDirectory.getAbsolutePath() + " to test classpath");
 			project.addTestCompileSourceRoot(target.outputDirectory.getAbsolutePath());
 		}
 		if (mainAddSources || testAddSources) {
@@ -400,6 +500,7 @@ public class ProtocJarMojo extends AbstractMojo
 			}
 
 			if (pluginPath != null) {
+				getLog().info("    Plugin path: " + pluginPath);
 				cmd.add("--plugin=protoc-gen-" + type + "=" + pluginPath);
 			}
 		}
@@ -413,6 +514,23 @@ public class ProtocJarMojo extends AbstractMojo
 			if (!include.exists()) throw new MojoExecutionException("Include path '" + include.getPath() + "' does not exist");
 			if (!include.isDirectory()) throw new MojoExecutionException("Include path '" + include.getPath() + "' is not a directory");
 			args.add("-I" + include.getPath());
+		}
+	}
+
+	private File resolveArtifact(String artifactSpec) throws MojoExecutionException {
+		try {
+			Properties detectorProps = new Properties();
+			new PlatformDetector().doDetect(detectorProps);
+			String platform = detectorProps.getProperty("os.detected.classifier");
+			
+			getLog().info("Resolving artifact: " + artifactSpec + ", platform: " + platform);
+			String[] as = artifactSpec.split(":");
+			Artifact artifact = artifactFactory.createDependencyArtifact(as[0], as[1], VersionRange.createFromVersionSpec(as[2]), "exe", platform, Artifact.SCOPE_RUNTIME);
+			artifactResolver.resolve(artifact, remoteRepositories, localRepository);
+			return artifact.getFile();
+		}
+		catch (Exception e) {
+			throw new MojoExecutionException("Error resolving artifact: " + artifactSpec, e);
 		}
 	}
 
@@ -430,6 +548,23 @@ public class ProtocJarMojo extends AbstractMojo
 
 		public boolean accept(File file) {
 			return file.getName().endsWith(extension);
+		}
+	}
+
+	class PlatformDetector extends Detector
+	{
+		void doDetect(Properties props) {
+	    	detect(props, null);
+		}
+
+		@Override
+		protected void log(String msg) {
+			//System.out.println(msg);
+		}
+
+		@Override
+		protected void logProperty(String name, String value) {
+			//log(name + ": " + value);
 		}
 	}
 }